Server/requirements.txt
Flask
flask-cors
python-dotenv

langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters

sentence-transformers
chromadb
groq

huggingface-hub
torch
transformers

unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
unstructured[pdf]
unstructured[html]
unstructured[image]
unstructured[video]
unstructured[docx]
unstructured[ocr]
unstructured[embeddings]

requests
beautifulsoup4

numpy
langchain-groq

